TestModule/requirements.txt
antlr4-python3-runtime==4.9.3
arrow==1.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock
fonttools==4.54.1
fsspec==2024.10.0
gmpy2 
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2 
kiwisolver==1.4.7
lazy_loader==0.4
loguru==0.7.2
MarkupSafe 
matplotlib==3.9.2
mpmath 
networkx 
numpy==2.1.3
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML 
requests==2.32.3
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
types-python-dateutil==2.9.0.20241003
typing_extensions
tzdata==2024.2
urllib3==2.2.3
win32-setctime==1.1.0
zipp==3.21.0
openpyxl
